Move ReadOptimizedLruCache
