+++ hazelcast/src/main/java/com/hazelcast/sql/impl/client/SqlClientService.java
@@ -33,6 +33,7 @@ import com.hazelcast.core.OperationTimeoutException;
 import com.hazelcast.internal.nio.ConnectionType;
 import com.hazelcast.internal.serialization.Data;
 import com.hazelcast.internal.serialization.InternalSerializationService;
+import com.hazelcast.internal.util.collection.ReadOptimizedLruCache;
 import com.hazelcast.logging.ILogger;
 import com.hazelcast.sql.HazelcastSqlException;
 import com.hazelcast.sql.SqlResult;
